Create Project-1-Explore-Weather-Trends
This project aims to extract the data from the SQL server and then calculating the Temperature moving average for my city and compare it with the global temperature. 

Special Note: 
Riyadh data starts from 1843 but there are two missing columns in 1846 and 1847. So, I started to apply the moving average method starting from 1848. Furthermore, to make a correct comparison I also started to apply the moving average method for the global data starting from 1848 until 2004. I had made every ten years into one interval to make it better to how the temperature is changing.